+++ TimberCommons/Settings/Configurator.cs
@@ -15,6 +15,7 @@ sealed class Configurator : IConfigurator {
     containerDefinition.Bind<TimeAndDurationSettings>().AsSingleton();
     containerDefinition.Bind<IrrigationSystemSettings>().AsSingleton();
     containerDefinition.Bind<WaterBuildingsSettings>().AsSingleton();
+    containerDefinition.Bind<InjuryProbabilitySettings>().AsSingleton();
     containerDefinition.Bind<DebugSettings>().AsSingleton();
   }
 }