+++ TimberCommons/CHANGELOG.md
@@ -1,4 +1,5 @@
 # v1.9.0 (started on 9/26/2024):
+* [Feature] Show injury probability on the buildings that can cause injury.
 
 # v1.8.3 (25 Sep 2024):
 * Migrate to Update 6. Not compatible with the previous versions.